requirements.txt
#LangChain Core
langchain
langchain_core

#OpenAI Integration
langchain_openai
openai

#Anthropic Integration
langchain_anthropic

#Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

#Hugging Face Integration
langchain_huggingface
transformers
huggingface-hub

#Environment Variables Management
python-dotenv

#Machine Learning Utilities
numpy
scikit-learn
streamlit
sentence-transformers